Refactor discovery and MCP services out of main
