+++ app/discovery/__init__.py
@@ -1,0 +1,31 @@
+from .service import (
+    capabilities_payload,
+    contracts_payload,
+    discovery_payload,
+    discovery_tools_payload,
+    discovery_workflows_payload,
+    handle_mcp_rpc_request,
+    health_payload,
+    manifest_payload,
+    tool_catalog,
+    tool_schema_lookup,
+    well_known_cognirelay_payload,
+    well_known_mcp_payload,
+    workflow_catalog,
+)
+
+__all__ = [
+    "capabilities_payload",
+    "contracts_payload",
+    "discovery_payload",
+    "discovery_tools_payload",
+    "discovery_workflows_payload",
+    "handle_mcp_rpc_request",
+    "health_payload",
+    "manifest_payload",
+    "tool_catalog",
+    "tool_schema_lookup",
+    "well_known_cognirelay_payload",
+    "well_known_mcp_payload",
+    "workflow_catalog",
+]
